passing down api data
