+++ app/components/IndividualPost.js
@@ -2,8 +2,6 @@ var React = require('react');
 var ReactDOM = require('react-dom');
 var ReactRouter = require('react-router');
 var Router = ReactRouter.Router;
-var RouteHandler = Router.RouteHandler;
-
 var Link = ReactRouter.Link;
